# go-server-server/go/model_article.go
/*
 * Blog API
 *
 * This is a mini blog.
 *
 * API version: 1.0.0
 * Contact: chris-ju@qq.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Article struct {

	Id int32 `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	Content string `json:"content,omitempty"`

	Date string `json:"date,omitempty"`

	Tag []Tag `json:"tag,omitempty"`

	User *User `json:"user,omitempty"`

	Review []Review `json:"review,omitempty"`
}