# go-server-server/go/model_review.go
/*
 * Blog API
 *
 * This is a mini blog.
 *
 * API version: 1.0.0
 * Contact: chris-ju@qq.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Review struct {

	Id int32 `json:"id,omitempty"`

	Userid int32 `json:"userid,omitempty"`

	Articleid int32 `json:"articleid,omitempty"`

	Content string `json:"content,omitempty"`
}