# go-server-server/go/model_tag.go
/*
 * Blog API
 *
 * This is a mini blog.
 *
 * API version: 1.0.0
 * Contact: chris-ju@qq.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Tag struct {

	Id int32 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`
}