# go-server-server/go/model_user.go
/*
 * Blog API
 *
 * This is a mini blog.
 *
 * API version: 1.0.0
 * Contact: chris-ju@qq.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type User struct {

	Id int32 `json:"id,omitempty"`

	Username string `json:"username,omitempty"`

	Nickname string `json:"nickname,omitempty"`

	Introduction string `json:"introduction,omitempty"`

	Email string `json:"email,omitempty"`

	Password string `json:"password,omitempty"`
}