# go-server-server/go/model_user_login.go
/*
 * Blog API
 *
 * This is a mini blog.
 *
 * API version: 1.0.0
 * Contact: chris-ju@qq.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type UserLogin struct {

	Username string `json:"username,omitempty"`

	Password string `json:"password,omitempty"`
}